Beta version .1. First working version. It now calculates Covered Call and Cash Secured Put Returns!

+++ Controllers/ThetaReturnCalculatorController.cs
@@ -10,7 +10,7 @@ namespace StockOptionsHelper.Controllers {
 			CoveredCall, CashSecuredPut
 		}
 
-		public void CalculateCoveredCallReturn(DateTime expirationDate, decimal sharePrice, decimal strike, int quantity, decimal contractPrice,
+		public void CalculateCoveredCallReturn(DateTime expirationDate, decimal sharePrice, decimal strike, decimal quantity, decimal contractPrice,
 			out decimal investedAmount, out decimal expectedProfit, out decimal expectedProfitPerc, out decimal expectedProfitAnnualPerc,
 			out decimal expectedMaxProfit, out decimal expectedMaxProfitPerc, out decimal expectedMaxProfitAnnualPerc, int OPTION_SHARES_PER_CONTRACT = 100) {
 
@@ -23,7 +23,7 @@ namespace StockOptionsHelper.Controllers {
 			expectedMaxProfitAnnualPerc = CalculateAnnualPerc(expirationDate, investedAmount, expectedMaxProfit);
 		}
 
-		public void CalcualteCashSecuredPutReturn(DateTime expirationDate, decimal sharePrice, decimal strike, int quantity, decimal contractPrice,
+		public void CalcualteCashSecuredPutReturn(DateTime expirationDate, decimal sharePrice, decimal strike, decimal quantity, decimal contractPrice,
 			out decimal investedAmount, out decimal expectedProfit, out decimal expectedProfitPerc, out decimal expectedProfitAnnualPerc,
 			out decimal expectedMaxProfit, out decimal expectedMaxProfitPerc, out decimal expectedMaxProfitAnnualPerc, int OPTION_SHARES_PER_CONTRACT = 100) {
